--- aspnet-core/modules/mvc/LINGYUN.Abp.AspNetCore.Mvc.Wrapper/LINGYUN/Abp/AspNetCore/Mvc/Wrapper/Wraping/IActionResultWrapper.cs
@@ -1,9 +1,0 @@
-﻿using Microsoft.AspNetCore.Mvc.Filters;
-
-namespace LINGYUN.Abp.AspNetCore.Mvc.Wrapper.Wraping
-{
-    public interface IActionResultWrapper
-    {
-        void Wrap(FilterContext context);
-    }
-}
