--- aspnet-core/modules/mvc/LINGYUN.Abp.AspNetCore.Mvc.Wrapper/LINGYUN/Abp/AspNetCore/Mvc/Wrapper/Wraping/IActionResultWrapperFactory.cs
@@ -1,10 +1,0 @@
-﻿using Microsoft.AspNetCore.Mvc.Filters;
-using Volo.Abp.DependencyInjection;
-
-namespace LINGYUN.Abp.AspNetCore.Mvc.Wrapper.Wraping
-{
-    public interface IActionResultWrapperFactory : ITransientDependency
-    {
-        IActionResultWrapper CreateFor(FilterContext context);
-    }
-}
