--- aspnet-core/modules/mvc/LINGYUN.Abp.AspNetCore.Mvc.Wrapper/LINGYUN/Abp/AspNetCore/Mvc/Wrapper/Wraping/NullActionResultWrapper.cs
@@ -1,12 +1,0 @@
-﻿using Microsoft.AspNetCore.Mvc.Filters;
-
-namespace LINGYUN.Abp.AspNetCore.Mvc.Wrapper.Wraping
-{
-    public class NullActionResultWrapper : IActionResultWrapper
-    {
-        public void Wrap(FilterContext context)
-        {
-
-        }
-    }
-}
