+++ aspnet-core/modules/mvc/LINGYUN.Abp.AspNetCore.Mvc.Wrapper/LINGYUN/Abp/AspNetCore/Mvc/Wrapper/Wrapping/IActionResultWrapper.cs
@@ -1,0 +1,9 @@
+﻿using Microsoft.AspNetCore.Mvc.Filters;
+
+namespace LINGYUN.Abp.AspNetCore.Mvc.Wrapper.Wrapping
+{
+    public interface IActionResultWrapper
+    {
+        void Wrap(FilterContext context);
+    }
+}
