+++ aspnet-core/modules/mvc/LINGYUN.Abp.AspNetCore.Mvc.Wrapper/LINGYUN/Abp/AspNetCore/Mvc/Wrapper/Wrapping/IActionResultWrapperFactory.cs
@@ -1,0 +1,10 @@
+﻿using Microsoft.AspNetCore.Mvc.Filters;
+using Volo.Abp.DependencyInjection;
+
+namespace LINGYUN.Abp.AspNetCore.Mvc.Wrapper.Wrapping
+{
+    public interface IActionResultWrapperFactory : ITransientDependency
+    {
+        IActionResultWrapper CreateFor(FilterContext context);
+    }
+}
