+++ aspnet-core/modules/mvc/LINGYUN.Abp.AspNetCore.Mvc.Wrapper/LINGYUN/Abp/AspNetCore/Mvc/Wrapper/Wrapping/NullActionResultWrapper.cs
@@ -1,0 +1,12 @@
+﻿using Microsoft.AspNetCore.Mvc.Filters;
+
+namespace LINGYUN.Abp.AspNetCore.Mvc.Wrapper.Wrapping
+{
+    public class NullActionResultWrapper : IActionResultWrapper
+    {
+        public void Wrap(FilterContext context)
+        {
+
+        }
+    }
+}
